Add trace viewer
