+++ run-server.sh
@@ -1,10 +1,6 @@
 #!/bin/bash
 cd "$(dirname "$0")"
 
-pip install -r aworld/requirements.txt && \
-
-pip install -r mcp_servers/requirements.txt && \
-
-pip install -r examples/web/requirements.txt && \
+pip install -e . && \
 
 streamlit run examples/web/main.py --server.port 8000
